+++ src/get_constants.py
@@ -1,7 +1,11 @@
+"""This module defines how the robot should receive and store constants/credentials from OpenOrchestrator."""
+
 from OpenOrchestrator.orchestrator_connection.connection import OrchestratorConnection
 
-def get_constants(orchestrator_connection:OrchestratorConnection) -> dict:
+def get_constants(orchestrator_connection: OrchestratorConnection) -> dict:
     """Get all constants used by the robot."""
+    orchestrator_connection.log_trace("Getting constants.")
+
     constants = {}
 
     # Get email address to send error screenshots to
