+++ src/initialize.py
@@ -1,5 +1,7 @@
+"""This module defines any initial processes to run when the robot starts."""
+
 from OpenOrchestrator.orchestrator_connection.connection import OrchestratorConnection
 
-def initialize(orchestrator_connection:OrchestratorConnection) -> None:
+def initialize(orchestrator_connection: OrchestratorConnection) -> None:
     """Do all custom startup initializations of the robot."""
-    pass
+    orchestrator_connection.log_trace("Initializing.")
