+++ src/reset.py
@@ -1,3 +1,5 @@
+"""This module handles resetting the state of the computer so the robot can work with a clean slate."""
+
 from OpenOrchestrator.orchestrator_connection.connection import OrchestratorConnection
 from OpenOrchestrator.database.constants import Credential
 
@@ -14,19 +16,21 @@ def reset(orchestrator_connection:OrchestratorConnection, sap_credentials: Crede
 
 def clean_up(orchestrator_connection: OrchestratorConnection) -> None:
     """Do any cleanup needed to leave a blank slate."""
-    pass
+    orchestrator_connection.log_trace("Doing cleanup.")
 
 
 def close_all(orchestrator_connection: OrchestratorConnection) -> None:
     """Gracefully close all applications used by the robot."""
-    pass
+    orchestrator_connection.log_trace("Closing all applications.")
 
 
 def kill_all(orchestrator_connection: OrchestratorConnection) -> None:
     """Forcefully close all applications used by the robot."""
+    orchestrator_connection.log_trace("Killing all applications.")
     sap_login.kill_sap()
 
 
 def open_all(orchestrator_connection: OrchestratorConnection, sap_credentials: Credential) -> None:
     """Open all programs used by the robot."""
+    orchestrator_connection.log_trace("Opening all applications.")
     sap_login.login_using_cli(sap_credentials.username, sap_credentials.password)
