+++ src/sap/handle_7.py
@@ -1,11 +1,12 @@
 """This module handles all Rykkerspærre of type '7'."""
 
 from datetime import date, timedelta
+
 import pyodbc
-
 from OpenOrchestrator.orchestrator_connection.connection import OrchestratorConnection
-
 from itk_dev_shared_components.sap import gridview_util, opret_kundekontakt
+
+from . import common
 
 def handle_7(orchestrator_connection:OrchestratorConnection, session, fmcacov_session):
     """Go through the list of type '7' rykkerspærre and handle them."""
@@ -47,14 +48,7 @@ def extend_all_rykkerspaerrer_deadlines(session, row_indices):
     """Extend all deadlines on the rykkerspærrer on the given rows, assuming they have been handled."""
     case_table = session.findById('wnd[0]/usr/cntlGRID1/shellcont/shell')
     for row in row_indices:
-        case_table.doubleClick(row, 'ZZ_IDNUMBER')
-
-        # Open aftaleindhold
-        session.findById("wnd[0]/usr/tabsTABSTRIP/tabpBUTOBJ").select()
-        afklaring_table = session.findById('wnd[0]/usr/tabsTABSTRIP/tabpBUTOBJ/ssubTABSUB:SAPLEMMA_CASE_TRANSACTION:0210/cntlWORKAREA1/shellcont/shell')
-        row_index = gridview_util.find_row_index_by_value(afklaring_table, 'CELEMNAME', 'Aftaleindhold')
-        afklaring_table.setCurrentCell(row_index, 'ID')
-        afklaring_table.clickCurrentCell()
+        common.open_aftaleindhold(session, case_table, row)
 
         # Check rykkerspærre type
         if session.findById("wnd[0]/usr/subBDT_AREA:SAPLBUSS:0021/tabsBDT_TABSTRIP01/tabpBUSCR02_01/ssubGENSUB:SAPLBUSS:0029/ssubGENSUB:SAPLBUSS:7135/subA04P02:SAPLFMCA_PSOB_BDT2:0330/ctxtSPSOB_SCR_2110_H3-DUNN_REASON").text != '7':
