+++ src/sap/handle_q.py
@@ -2,21 +2,15 @@
 
 from itk_dev_shared_components.sap import gridview_util
 
+from . import common
+
 def handle_q(session):
     """Handle all Rykkerspærre of type 'Q'."""
     case_table = session.findById('wnd[0]/usr/cntlGRID1/shellcont/shell')
     gridview_util.scroll_entire_table(case_table, True)
 
     for row in range(case_table.RowCount):
-        # Open case
-        case_table.doubleClick(row, 'ZZ_IDNUMBER')
-
-        # Open aftaleindhold
-        session.findById("wnd[0]/usr/tabsTABSTRIP/tabpBUTOBJ").select()
-        afklaring_table = session.findById('wnd[0]/usr/tabsTABSTRIP/tabpBUTOBJ/ssubTABSUB:SAPLEMMA_CASE_TRANSACTION:0210/cntlWORKAREA1/shellcont/shell')
-        row_index = gridview_util.find_row_index_by_value(afklaring_table, 'CELEMNAME', 'Aftaleindhold')
-        afklaring_table.setCurrentCell(row_index, 'ID')
-        afklaring_table.clickCurrentCell()
+        common.open_aftaleindhold(session, case_table, row)
 
         # Press edit, delete 'Q' and save
         if session.findById("wnd[0]/usr/subBDT_AREA:SAPLBUSS:0021/tabsBDT_TABSTRIP01/tabpBUSCR02_01/ssubGENSUB:SAPLBUSS:0029/ssubGENSUB:SAPLBUSS:7135/subA04P02:SAPLFMCA_PSOB_BDT2:0330/ctxtSPSOB_SCR_2110_H3-DUNN_REASON").text == 'Q':
